Restructure project to enable adding more constraints more easily. All check-specific code now lives in types/check.rb
